config: replace user and repo fields with one repo
Also fix a bug in shutdown. Before this commit, if one of the
async threads paniced the main thread would hang. The solution
is to not join explictly one at a time, but let the scope thread
thing do the joining. This means that all shutdown signals run
before joining on threads, and we don't get stuck joining.